LayoutTests/imported/w3c/web-platform-tests/workers/Worker_ErrorEvent_bubbles_cancelable-expected.txt

CONSOLE MESSAGE: Error: Error Message

PASS ErrorEvent on worker doesn't bubble and is cancelable
